Move dict_get tests
