+++ test/test_traversal.py
@@ -3,7 +3,7 @@ import xml.etree.ElementTree
 
 import pytest
 
-from yt_dlp.utils import int_or_none, str_or_none
+from yt_dlp.utils import dict_get, int_or_none, str_or_none
 from yt_dlp.utils.traversal import traverse_obj
 
 _TEST_DATA = {
@@ -25,6 +25,27 @@ _TEST_DATA = {
 
 
 class TestTraversal:
+    def test_dict_get(self):
+        FALSE_VALUES = {
+            'none': None,
+            'false': False,
+            'zero': 0,
+            'empty_string': '',
+            'empty_list': [],
+        }
+        d = {**FALSE_VALUES, 'a': 42}
+        assert dict_get(d, 'a') == 42
+        assert dict_get(d, 'b') == None
+        assert dict_get(d, 'b', 42) == 42
+        assert dict_get(d, ('a',)) == 42
+        assert dict_get(d, ('b', 'a')) == 42
+        assert dict_get(d, ('b', 'c', 'a', 'd')) == 42
+        assert dict_get(d, ('b', 'c')) == None
+        assert dict_get(d, ('b', 'c'), 42) == 42
+        for key, false_value in FALSE_VALUES.items():
+            assert dict_get(d, ('b', 'c', key)) == None
+            assert dict_get(d, ('b', 'c', key), skip_false_values=False) == false_value
+
     def test_traversal_base(self):
         assert traverse_obj(_TEST_DATA, ('str',)) == 'str', \
             'allow tuple path'
